Refactoring Input
